requirements.txt
Flask==2.0.3
pytesseract==0.3.9
opencv-python==4.5.5.62
Pillow==9.0.1